refactor: interestService interface 추가
